requirements.txt
fastapi==0.68.1
alembic==1.8.1
gunicorn==20.1.0
fastapi-event==0.1.3
uvicorn==0.15.0
grpcio==1.40.0
grpcio-tools==1.40.0
pika==1.2.0
protobuf==3.18.0
pre-commit==3.3.3
flake8==4
black==23.7.0
python-json-logger==2.0.7
python-dotenv==1.0.0
